Start moving shared classes and resources to common
