+++ android/common/src/main/java/com/novoda/peepz/PictureTakeListener.java
