+++ android/common/src/main/res/drawable-hdpi/ic_camera_white.png
