+++ android/common/src/main/res/drawable-mdpi/ic_camera_white.png
