+++ android/common/src/main/res/drawable-xhdpi/ic_camera_white.png
