+++ android/common/src/main/res/drawable-xxhdpi/ic_camera_white.png
