+++ android/common/src/main/res/drawable/purple_gradient_circle_background.xml
