+++ android/common/src/main/res/layout/merge_floating_action_button.xml
