+++ android/common/src/main/res/values/attrs.xml
