+++ android/dependencies.gradle
@@ -1,11 +1,12 @@
 ext {
     supportLibVersion = '27.1.1'
-    firebaseVersion = '10.0.1'
+    firebaseVersion = '12.0.1'
+    butterknifeVersion = '8.4.0'
 
     libraries = [
             accessibilitools    : 'com.novoda:accessibilitools:1.4.0',
-            butterknife         : 'com.jakewharton:butterknife:8.8.1',
-            butterknifeProcessor: 'com.jakewharton:butterknife-compiler:8.8.1',
+            butterknife         : "com.jakewharton:butterknife:${butterknifeVersion}",
+            butterknifeProcessor: "com.jakewharton:butterknife-compiler:${butterknifeVersion}",
             cameraview          : 'com.github.jackSzm:cameraview:e32e238321',
             findbugsAnnotations : 'com.google.code.findbugs:annotations:3.0.0',
             firebaseAuth        : "com.google.firebase:firebase-auth:${firebaseVersion}",
@@ -13,7 +14,7 @@ ext {
             firebaseDatabase    : "com.google.firebase:firebase-database:${firebaseVersion}",
             firebaseStorage     : "com.google.firebase:firebase-storage:${firebaseVersion}",
             glide               : 'com.github.bumptech.glide:glide:3.7.0',
-            playServicesAuth    : 'com.google.android.gms:play-services-auth:10.0.1',
+            playServicesAuth    : 'com.google.android.gms:play-services-auth:12.0.1',
             rvTools             : 'com.ataulm:rv-tools:1.1.0',
             supportAppCompat    : "com.android.support:appcompat-v7:${supportLibVersion}",
             supportDesign       : "com.android.support:design:${supportLibVersion}",
