+++ android/mobile/build.gradle
@@ -27,6 +27,7 @@ android {
 }
 
 dependencies {
+    implementation project(":core")
     implementation project(":common")
 
     implementation libraries.supportAppCompat
